refactored StudentDetails page
